+++ src/Hagar.CodeGenerator/Model/FieldDescription.cs
@@ -4,13 +4,13 @@ namespace Hagar.CodeGenerator
 {
     internal class FieldDescription : IFieldDescription
     {
-        public FieldDescription(uint fieldId, IFieldSymbol field)
+        public FieldDescription(ushort fieldId, IFieldSymbol field)
         {
             this.FieldId = fieldId;
             this.Field = field;
         }
         public IFieldSymbol Field { get; }
-        public uint FieldId { get; }
+        public ushort FieldId { get; }
         public ISymbol Member => this.Field;
         public ITypeSymbol Type => this.Field.Type;
         public string Name => this.Field.Name;
